requirements.txt
PyQt6
random
math
opencv-python
numpy
ultralytics
jkp-depth-anything-v2